+++ crates/goose/src/plugins/formats/mod.rs
@@ -1,0 +1,1 @@
+pub(super) mod gemini;
